+++ pandapower/toolbox.py
@@ -22,21 +22,13 @@ from itertools import chain
 import networkx as nx
 import numpy as np
 import pandas as pd
+
 from packaging import version
-from pandapower.auxiliary import _preserve_dtypes, get_indices, pandapowerNet
-from pandapower.create import (
-    create_bus,
-    create_empty_network,
-    create_ext_grid,
-    create_gen,
-    create_impedance,
-    create_line_from_parameters,
-    create_load,
-    create_sgen,
-    create_shunt,
-    create_storage,
-    create_switch,
-)
+import numbers
+from pandapower.auxiliary import get_indices, pandapowerNet, _preserve_dtypes
+from pandapower.create import create_switch, create_line_from_parameters, \
+    create_impedance, create_empty_network, create_gen, create_ext_grid, \
+    create_load, create_shunt, create_bus, create_sgen, create_storage
 from pandapower.opf.validate_opf_input import _check_necessary_opf_parameters
 from pandapower.run import runpp
 from pandapower.std_types import change_std_type
@@ -1066,12 +1058,8 @@ def _nets_equal_keys(
             if not isinstance(net2[key], np.ndarray):
                 not_equal.append(key)
             else:
-                if version.parse(np.__version__) < version.parse("0.19"):
-                    if not compare_arrays(net1[key], net2[key]).all():
-                        not_equal.append(key)
-                else:
-                    if not np.array_equal(net1[key], net2[key], equal_nan=True):
-                        not_equal.append(key)
+                if not np.array_equal(net1[key], net2[key], equal_nan=True):
+                    not_equal.append(key)
 
         elif (
             isinstance(net1[key], int)
@@ -1527,20 +1515,9 @@ def set_data_type_of_columns_to_default(net):
                     if new_net[key][col].dtype == net[key][col].dtype:
                         continue
                     if set(item.columns) == set(new_net[key]):
-                        if version.parse(pd.__version__) < version.parse("0.21"):
-                            net[key] = net[key].reindex_axis(
-                                new_net[key].columns, axis=1
-                            )
-                        else:
-                            net[key] = net[key].reindex(new_net[key].columns, axis=1)
-                    if version.parse(pd.__version__) < version.parse("0.20.0"):
-                        net[key][col] = net[key][col].astype(
-                            new_net[key][col].dtype, raise_on_error=False
-                        )
-                    else:
-                        net[key][col] = net[key][col].astype(
-                            new_net[key][col].dtype, errors="ignore"
-                        )
+                        net[key] = net[key].reindex(new_net[key].columns, axis=1)
+                    net[key][col] = net[key][col].astype(new_net[key][col].dtype,
+                                                         errors="ignore")
 
 
 # --- Modify topology
@@ -2833,10 +2810,7 @@ def replace_ext_grid_by_gen(
         in_res = pd.Series(ext_grids).isin(net["res_ext_grid"].index).values
         to_add = net.res_ext_grid.loc[pd.Index(ext_grids)[in_res]]
         to_add.index = pd.Index(new_idx)[in_res]
-        if version.parse(pd.__version__) < version.parse("0.23"):
-            net.res_gen = pd.concat([net.res_gen, to_add])
-        else:
-            net.res_gen = pd.concat([net.res_gen, to_add], sort=True)
+        net.res_gen = pd.concat([net.res_gen, to_add], sort=True)
         net.res_ext_grid.drop(pd.Index(ext_grids)[in_res], inplace=True)
     return new_idx
 
@@ -2941,10 +2915,7 @@ def replace_gen_by_ext_grid(
         in_res = pd.Series(gens).isin(net["res_gen"].index).values
         to_add = net.res_gen.loc[pd.Index(gens)[in_res]]
         to_add.index = pd.Index(new_idx)[in_res]
-        if version.parse(pd.__version__) < version.parse("0.23"):
-            net.res_ext_grid = pd.concat([net.res_ext_grid, to_add])
-        else:
-            net.res_ext_grid = pd.concat([net.res_ext_grid, to_add], sort=True)
+        net.res_ext_grid = pd.concat([net.res_ext_grid, to_add], sort=True)
         net.res_gen.drop(pd.Index(gens)[in_res], inplace=True)
     return new_idx
 
@@ -3054,10 +3025,7 @@ def replace_gen_by_sgen(
         in_res = pd.Series(gens).isin(net["res_gen"].index).values
         to_add = net.res_gen.loc[pd.Index(gens)[in_res]]
         to_add.index = pd.Index(new_idx)[in_res]
-        if version.parse(pd.__version__) < version.parse("0.23"):
-            net.res_sgen = pd.concat([net.res_sgen, to_add])
-        else:
-            net.res_sgen = pd.concat([net.res_sgen, to_add], sort=True)
+        net.res_sgen = pd.concat([net.res_sgen, to_add], sort=True)
         net.res_gen.drop(pd.Index(gens)[in_res], inplace=True)
     return new_idx
 
@@ -3180,10 +3148,7 @@ def replace_sgen_by_gen(
         in_res = pd.Series(sgens).isin(net["res_sgen"].index).values
         to_add = net.res_sgen.loc[pd.Index(sgens)[in_res]]
         to_add.index = pd.Index(new_idx)[in_res]
-        if version.parse(pd.__version__) < version.parse("0.23"):
-            net.res_gen = pd.concat([net.res_gen, to_add])
-        else:
-            net.res_gen = pd.concat([net.res_gen, to_add], sort=True)
+        net.res_gen = pd.concat([net.res_gen, to_add], sort=True)
         net.res_sgen.drop(pd.Index(sgens)[in_res], inplace=True)
     return new_idx
 
@@ -3341,13 +3306,13 @@ def replace_pq_elmtype(
         in_res = pd.Series(old_indices).isin(net["res_" + old_elm].index).values
         to_add = net["res_" + old_elm].loc[pd.Index(old_indices)[in_res]]
         to_add.index = pd.Index(new_idx)[in_res]
+
         if version.parse(pd.__version__) < version.parse("0.23"):
             net["res_" + new_elm] = pd.concat([net["res_" + new_elm], to_add])
         else:
             net["res_" + new_elm] = pd.concat(
                 [net["res_" + new_elm], to_add], sort=True
             )
-        net["res_" + old_elm].drop(pd.Index(old_indices)[in_res], inplace=True)
     return new_idx