+++ CHANGELOG.rst
@@ -7,6 +7,10 @@ Change Log
 - [FIX] from_ppc() converter and power system test cases: add missing factor for tap_side=="lv"; change tap_side to "hv" for all test cases (were converted without new factor, so as the tap_side is "hv")
 - [ADDED] from_mpc() converter: added functionality to import .m files via external package
 - [CHANGED] from_ppc() converter: added option of tap_side and essential speed up
+- [CHANGED] drop support of pandas versions < 1.0
+- [ADDED] parameter in_ka for rated switch current
+- [ADDED] current and loading result for switches
+- [FIXED] bug for disabled continous tap controllers
 
 [2.9.0]- 2022-03-23
 ----------------------